+++ sphinxcontrib/proof/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Louis Paternault
+# Copyright 2018 Louis Paternault
 #
 # Sphinxcontrib-Proof is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,24 +16,32 @@
 """Provide tools to typeset theorems, proofs, etc. in Sphinx documentation."""
 
 import os
-import pkg_resources
-
 from docutils import nodes
 from docutils.parsers.rst import directives, Directive
 
 from sphinx.util import copy_static_entry
 from sphinx.util.nodes import set_source_info
 
-
-def package_file(*filename):
-    """Return the path to a filename present in package data."""
-    return pkg_resources.resource_filename(
-        "sphinxcontrib.proof", os.path.join("data", *filename)
-    )
-
-
 VERSION = "1.0.1"
 PREFIX = "proof:"
+
+################################################################################
+# Configuration
+PROOF_THEOREM_LABELS = {
+    "algorithm": "Algorithm",
+    "conjecture": "Conjecture",
+    "corollary": "Corollary",
+    "definition": "Definition",
+    "example": "Example",
+    "lemma": "Lemma",
+    "observation": "Observation",
+    "proof": "Proof",
+    "property": "Property",
+    "theorem": "Theorem",
+}
+
+################################################################################
+# Docutils
 
 
 class StatementNode(nodes.General, nodes.Element):
@@ -42,59 +50,10 @@ class StatementNode(nodes.General, nodes.Element):
     pass
 
 
-class ProofNode(nodes.Part, nodes.Element):
-    """Proof"""
-
-    pass
-
-
 class ContentNode(nodes.General, nodes.Element):
     """Content of a proof or a statement"""
 
     pass
-
-
-# This should be internationalized using gettext… Patch welcome!
-FRENCH = {
-    "lemma": u"Lemme",
-    "property": u"Propriété",
-    "example": u"Exemple",
-    "theorem": u"Théorème",
-    "definition": u"Définition",
-    "proof": u"Preuve",
-    "conjecture": u"Conjecture",
-    "algorithm": u"Algorithme",
-}
-
-
-class ProofEnvironment(Directive):
-    """A proof environment"""
-
-    has_content = True
-    required_arguments = 0
-    optional_arguments = 1
-    final_argument_whitespace = True
-    option_spec = {"label": directives.unchanged_required}
-
-    def run(self):
-        """Render this environment"""
-        env = self.state.document.settings.env
-        targetid = "index-%s" % env.new_serialno("index")
-        targetnode = nodes.target("", "", ids=[targetid])
-
-        node = ProofNode("\n".join(self.content))
-        node["classes"] += ["proof-type-proof"]
-
-        if self.arguments:
-            node["title"] = self.arguments[0]
-
-        content = ContentNode()
-        self.state.nested_parse(self.content, self.content_offset, content)
-        content["classes"] += ["proof-content"]
-        node += content
-
-        set_source_info(self, node)
-        return [targetnode, node]
 
 
 class StatementEnvironment(Directive):
@@ -127,31 +86,17 @@ class StatementEnvironment(Directive):
         return [targetnode, node]
 
 
+################################################################################
 # HTML
-def html_visit_proof_node(self, node):
-    """Enter :class:`ProofNode` in HTML builder."""
-    self.body.append(self.starttag(node, "div"))
-    self.body.append("""<div class="proof-title">""")
-    self.body.append("""<span class="proof-title-name">Preuve</span>""")
-    if "title" in node:
-        self.body.append("""<span class="proof-title-content">""")
-        self.body.append(u"({})".format(node["title"]))
-        self.body.append("""</span>""")
-    self.body.append("""</div>""")
-
-
-def html_depart_proof_node(self, node):
-    """Leave :class:`ProofNode` in HTML builder."""
-    # pylint: disable=unused-argument
-    self.body.append("</div>")
-
-
 def html_visit_statement_node(self, node):
     """Enter :class:`StatementNode` in HTML builder."""
+    labels = self.builder.config.proof_theorem_labels
     self.body.append(self.starttag(node, "div"))
     self.body.append("""<div class="proof-title">""")
     self.body.append(
-        u"""<span class="proof-title-name">{}</span>""".format(FRENCH[node["name"]])
+        u"""<span class="proof-title-name">{}</span>""".format(
+            labels.get(node["name"], "unknown-theorem-type")
+        )
     )
     if "title" in node:
         self.body.append("""<span class="proof-title-content">""")
@@ -177,20 +122,8 @@ def html_depart_content_node(self, node):
     self.body.append("</div>")
 
 
+################################################################################
 # LaTeX
-def latex_visit_proof_node(self, node):
-    """Enter :class:`ProofNode` in LaTeX builder."""
-    self.body.append(r"\begin{proof}")
-    if "title" in node:
-        self.body.append("[{}]".format(node["title"]))
-    self.body.append("\n")
-
-
-def latex_depart_proof_node(self, node):
-    """Leave :class:`ProofNode` in LaTeX builder."""
-    # pylint: disable=unused-argument
-    self.body.append(r"\end{proof}")
-    self.body.append("\n")
 
 
 def latex_visit_statement_node(self, node):
@@ -219,13 +152,52 @@ def latex_depart_content_node(self, node):
     pass
 
 
+def _latex_preamble_iterator(config):
+    labels = config.proof_theorem_labels
+
+    yield config.latex_elements.get("preamble", "")
+
+    yield r"\makeatletter"
+
+    if config.latex_proof_counter:
+        yield r"\newtheorem{%s}{%s}[%s]" % (
+            config.latex_proof_main,
+            labels[config.latex_proof_main],
+            config.latex_proof_counter,
+        )
+    else:
+        yield r"\newtheorem{%s}{%s}" % (
+            config.latex_proof_main,
+            labels[config.latex_proof_main],
+        )
+
+    for environment, label in labels.items():
+        if not environment in config.latex_proof_notheorem + [config.latex_proof_main]:
+            yield r"\newtheorem{%s}[%s]{%s}" % (
+                environment,
+                config.latex_proof_main,
+                label,
+            )
+
+    yield r"\makeatother"
+
+
+def latex_preamble(config):
+    """Return the custom LaTeX preamble."""
+    return "\n".join(_latex_preamble_iterator(config))
+
+
+################################################################################
+# Setup
+
+
 def builder_inited(app):
     """Hook called when builder has been inited."""
+    config = app.builder.config
     if app.builder.name == "latex":
-        app.builder.config.latex_additional_files.append(
-            package_file("_static", "sphinxcontribproof.sty")
-        )
-        app.add_latex_package("sphinxcontribproof")
+        if "preamble" not in config.latex_elements:
+            config.latex_elements["preamble"] = ""
+        config.latex_elements["preamble"] += latex_preamble(config)
 
 
 def setup(app):
@@ -233,11 +205,11 @@ def setup(app):
     app.add_stylesheet("proof.css")
     app.add_javascript("proof.js")
 
-    app.add_node(
-        ProofNode,
-        html=(html_visit_proof_node, html_depart_proof_node),
-        latex=(latex_visit_proof_node, latex_depart_proof_node),
-    )
+    app.add_config_value("proof_theorem_labels", PROOF_THEOREM_LABELS, "env")
+    app.add_config_value("latex_proof_counter", "", "env")
+    app.add_config_value("latex_proof_main", "theorem", "env")
+    app.add_config_value("latex_proof_notheorem", [], "env")
+
     app.add_node(
         StatementNode,
         html=(html_visit_statement_node, html_depart_statement_node),
@@ -249,13 +221,7 @@ def setup(app):
         latex=(latex_visit_content_node, latex_depart_content_node),
     )
 
-    app.add_directive(PREFIX + "property", StatementEnvironment)
-    app.add_directive(PREFIX + "lemma", StatementEnvironment)
-    app.add_directive(PREFIX + "example", StatementEnvironment)
-    app.add_directive(PREFIX + "theorem", StatementEnvironment)
-    app.add_directive(PREFIX + "definition", StatementEnvironment)
-    app.add_directive(PREFIX + "proof", ProofEnvironment)
-    app.add_directive(PREFIX + "conjecture", StatementEnvironment)
-    app.add_directive(PREFIX + "algorithm", StatementEnvironment)
+    for environment in app.config.proof_theorem_labels:
+        app.add_directive(PREFIX + environment, StatementEnvironment)
 
     app.connect("builder-inited", builder_inited)